backend/pyproject.toml
[project]
name = "steam-collector"
version = "1.0.0"
description = "Steam 游戏时长数据采集器"
requires-python = ">=3.10"

dependencies = [
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "undetected-chromedriver>=3.5.5",
    "setuptools>=80.9.0",
]

[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

